Making part2 python3 friendly

+++ engine.py
@@ -17,8 +17,8 @@ def main():
         'dark_ground': libtcod.Color(50, 50, 150)
     }
 
-    player = Entity(screen_width / 2, screen_height / 2, '@', libtcod.white)
-    npc = Entity(screen_width / 2 - 5, screen_height / 2, '@', libtcod.yellow)
+    player = Entity(int(screen_width / 2), int(screen_height / 2), '@', libtcod.white)
+    npc = Entity(int(screen_width / 2 - 5), int(screen_height / 2), '@', libtcod.yellow)
     entities = [npc, player]
 
     libtcod.console_set_custom_font('arial10x10.png', libtcod.FONT_TYPE_GREYSCALE | libtcod.FONT_LAYOUT_TCOD)
